feat: Move customer phone number field into separate module
